analyzer/requirements.txt
requests
paho-mqtt
influxdb-client